feat: restructure authentication components and implement email verification notice
